Refactor
* move code in Store to AWSConfig
* rename Profile to ConfigEntry
* cleanup Parser code
* cleanup ProfileResolver

More details:

The code in Store was used only in AWSConfig file. This added
unnecessary indirection, since AWSConfig is an entry point for
the gem it's natural to have the code live in the file.

Profile has been renamed to ConfigEntry. It's representation of a
config. It is associated with a profile (accessed via AWSConfig),
but it isn't always representing a complete configuration. The class
is used to encapsulate nested properties too.

ProfileResolver always refreshes all sourced profiles. It makes the
code easier, but also avoids potential edge cases when a duplicate
profile is added, it's being merged and it should be re-sourced. This
it very unlikely to ever happen, but we don't need to worry about it
since we always re-source all profiles when anything is added.